Support for updating tasks added. Adding precedent task to the creation script.

+++ src/com/ipm/web/impl/manager/ResourceManagerImpl.java
@@ -16,7 +16,7 @@ public class ResourceManagerImpl implements ResourceManager {
 	private ResourceDao resourceDao;
 
 	@Override
-	public List<Resource> getResources(String username, int projectId) {
+	public List<Resource> getResources(String username, long projectId) {
 		return resourceDao.getResources(username,projectId);
 	}
